get rid of warnings, code cleanup and prepare for cuda kernel impl

+++ src/random.h
@@ -52,10 +52,10 @@ inline float fastrand(RandomCtx ctx)
     ctx.state = oldstate * 6364136223846793005ULL + (ctx.inc | 1);
     
     // Calculate output function (XSH RR), uses old state for max ILP
-    uint32_t xorshifted = ((oldstate >> 18u) ^ oldstate) >> 27u;
+    uint32_t xorshifted = static_cast<uint32_t>(((oldstate >> 18u) ^ oldstate) >> 27u);
     uint32_t rot = oldstate >> 59u;
 
-    uint32_t rand = (xorshifted >> rot) | (xorshifted << ((-rot) & 31));
+    uint32_t rand = (xorshifted >> rot) | (xorshifted << (-(static_cast<int32_t>(rot)) & 31));
 
     return rand / static_cast<float>(UINT32_MAX);
 }
